Separated module definition from simulation functions
